+++ pages/projects/digital-centre.js
@@ -77,9 +77,12 @@ export default function DigitalCenter(props) {
 
           {/* Primary HTML Meta Tags */}
           <title>{`${t("dc:OverviewTitle")} — ${t("siteTitle")}`}</title>
-          <meta name="description" content={`${t("dc:ProductGoal1")}`} />
+          <meta name="description" content={`${t("dc:MetaDescription")}`} />
           <meta name="author" content="Service Canada" />
           <link rel="icon" href="/favicon.ico" />
+          <link rel="schema.dcterms" href="http://purl.org/dc/terms/" />
+          <link rel="canonical" href={`${t("dc:CanonicalURL")}`} />
+          <meta name="keywords" content={`${t("dc:Keywords")}`} />
 
           {/* DCMI Meta Tags */}
           <meta
@@ -89,54 +92,53 @@ export default function DigitalCenter(props) {
           <meta
             name="dcterms.language"
             content={props.locale === "en" ? "eng" : "fra"}
+            title="ISO639-2/T"
           />
-          <meta name="dcterms.creator" content={t("creator")} />
+          <meta
+            name="dcterms.description"
+            content={`${t("dc:MetaDescription")}`}
+          />
+          <meta
+            name="dcterms.subject"
+            title="gccore"
+            content={`${t("dc:MetaSubject")}`}
+          />
+          <meta name="dcterms.creator" content="Service Canada" />
           <meta name="dcterms.accessRights" content="2" />
           <meta
             name="dcterms.service"
             content="ESDC-EDSC_SCLabs-LaboratoireSC"
           />
-          <meta name="dcterms.issued" content="2021-07-21" />
+          <meta name="dcterms.issued" title="W3CDTF" content="2021-07-21" />
+          <meta name="dcterms.modified" title="W3CDTF" content="2021-12-16" />
+          <meta name="dcterms.spatial" content="Canada" />
 
           {/* Open Graph / Facebook */}
           <meta property="og:type" content="website" />
           <meta property="og:locale" content={props.locale} />
-          <meta
-            property="og:url"
-            content={
-              "https://alpha.service.canada.ca/" +
-              `${props.locale}` +
-              `${t("projectRedirect")}` +
-              "/digital-centre"
-            }
-          />
+          <meta property="og:url" content={`${t("dc:CanonicalURL")}`} />
           <meta
             property="og:title"
             content={`${t("dc:OverviewTitle")} — ${t("siteTitle")}`}
           />
-          <meta property="og:description" content={`${t("dc:ProductGoal1")}`} />
+          <meta
+            property="og:description"
+            content={`${t("dc:MetaDescription")}`}
+          />
           <meta property="og:image" content={`${t("metaImage")}`} />
           <meta property="og:image:alt" content={`${t("siteTitle")}`} />
 
           {/* Twitter */}
           <meta property="twitter:card" content="summary_large_image" />
-          <meta
-            property="twitter:url"
-            content={
-              "https://alpha.service.canada.ca/" +
-              `${props.locale}` +
-              `${t("projectRedirect")}` +
-              "/digital-centre"
-            }
-          />
+          <meta property="twitter:url" content={`${t("dc:CanonicalURL")}`} />
           <meta
             property="twitter:title"
             content={`${t("dc:OverviewTitle")} — ${t("siteTitle")}`}
           />
-          <meta name="twitter:creator" content={t("creator")} />
+          <meta name="twitter:creator" content="Service Canada" />
           <meta
             property="twitter:description"
-            content={`${t("dc:ProductGoal1")}`}
+            content={`${t("dc:MetaDescription")}`}
           />
           <meta property="twitter:image" content={`${t("metaImage")}`} />
           <meta property="twitter:image:alt" content={`${t("siteTitle")}`} />